+++ src/nlp/similarity/simUtils.py
@@ -1,6 +1,6 @@
-import sys
 import math
 import numpy as np
+import copy
 # https://github.com/alvations/pywsd
 from pywsd.lesk import simple_lesk, original_lesk, cosine_lesk, adapted_lesk
 from pywsd import disambiguate
@@ -10,7 +10,9 @@ import nltk
 from nltk import word_tokenize as tokenizer
 from nltk.corpus import brown
 from nltk.corpus import wordnet as wn
-from nltk.corpus import wordnet_ic
+
+from synsetUtils import semanticSimilaritySynsets, synsetsSimilarity
+from synsetUtils import semanticSimilarityUsingDisambiguatedSynsets
 
 """
   Methods proposed by: https://ieeexplore.ieee.org/stamp/stamp.jsp?tp=&arnumber=1644735
@@ -195,18 +197,6 @@ def semanticSimilarityWords(wordA, wordB):
   similarity = semanticSimilaritySynsets(bestPair[0], bestPair[1], disambiguation=False)
   return similarity
 
-def semanticSimilaritySynsets(synsetA, synsetB, disambiguation=False):
-  """
-    Compute the similarity between two synset using semantic analysis
-    e.g., using both path length and depth information in wordnet
-    @ In, synsetA, wordnet.synset, the first synset
-    @ In, synsetB, wordnet.synset, the second synset
-    @ Out, similarity, float, [0, 1], the similarity score
-  """
-  shortDistance = pathLength(synsetA, synsetB, disambiguation=disambiguation)
-  maxHierarchy = scalingDepthEffect(synsetA, synsetB, disambiguation=disambiguation)
-  similarity = shortDistance*maxHierarchy
-  return similarity
 
 def identifyBestSimilarSynsetPair(wordA, wordB):
   """
@@ -232,100 +222,6 @@ def identifyBestSimilarSynsetPair(wordA, wordB):
           similarity = temp
           bestPair = synsetWordA, synsetWordB
     return bestPair
-
-def pathLength(synsetA, synsetB, alpha=0.2, disambiguation=False):
-  """
-    Path length calculation using nonlinear transfer function between two Wordnet Synsets
-    The two Synsets should be the best Synset Pair (e.g., disambiguation should be performed)
-    @ In, synsetA, wordnet.synset, synset for first word
-    @ In, synsetB, wordnet.synset, synset for second word
-    @ In, alpha, float, a constant in monotonically descreasing function, exp(-alpha*wordnetpathLength),
-      parameter used to scale the shortest path length. For wordnet, the optimal value is 0.2
-    @ In, disambiguation, bool, True if disambiguation have been performed for the given synsets
-    @ Out, shortDistance, float, [0, 1], the shortest distance between two synsets using exponential descreasing
-      function.
-  """
-  synsetA = wn.synset(synsetA.name())
-  synsetB = wn.synset(synsetB.name())
-  maxLength = sys.maxsize
-  if synsetA is None or synsetB is None:
-    return 0.0
-  # ? The original word is difference, but their synset are the same, we assume the path length is zero
-  if synsetA == synsetB:
-    maxLength = 0.0
-  else:
-    if not disambiguation:
-      lemmaSetA = set([str(word.name()) for word in synsetA.lemmas()])
-      lemmaSetB = set([str(word.name()) for word in synsetB.lemmas()])
-      #this line if the word is none
-      if len(lemmaSetA.intersection(lemmaSetB)) > 0:
-        maxLength = 1.0
-      else:
-        maxLength = synsetA.shortest_path_distance(synsetB)
-        if maxLength is None:
-          maxLength = 0.0
-    else:
-      # when disamigutation is performed, we should avoid to check lemmas
-      maxLength = synsetA.shortest_path_distance(synsetB)
-      if maxLength is None:
-        maxLength = 0.0
-  shortDistance = math.exp(-alpha*maxLength)
-  return shortDistance
-
-def scalingDepthEffect(synsetA, synsetB, beta=0.45, disambiguation=False):
-  """
-    Words at upper layers of hierarchical semantic nets have more general concepts and less semantic similarity
-    between words than words at lower layers. This method is used to scale the similarity behavior with repect
-    to depth h, e.g., [exp(beta*h)-exp(-beta*g)]/[exp(beta*h)+exp(-beta*g)]
-    The two Synsets should be the best Synset Pair (e.g., disambiguation should be performed)
-    @ In, synsetA, wordnet.synset, synset for first word
-    @ In, synsetB, wordnet.synset, synset for second word
-    @ In, beta, float, parameter used to scale the shortest depth, for wordnet, the optimal value is 0.45
-    @ In, disambiguation, bool, True if disambiguation have been performed for the given synsets
-    @ out, treeDist, float, [0, 1], similary score based on depth effect in wordnet
-  """
-  maxLength = sys.maxsize
-  smoothingFactor = beta
-  if synsetA is None or synsetB is None:
-    return 0.0
-
-  if synsetA == synsetB:
-    if disambiguation:
-      return 1.0
-    else:
-      # The following is from original code, I think it should be return 1.0 when synset are the same
-      # I think a new similarity calculations should be proposed
-      # values for different lengths and their similarity score:
-      # len simScore
-      # 0   0.0
-      # 1   0.4218990052500079
-      # 2   0.7162978701990244
-      # 3   0.874053287886007
-      # 4   0.9468060128462682
-      # 5   0.9780261147388136
-      # 6   0.9910074536781175
-      maxLength = max(word[1] for word in synsetA.hypernym_distances())
-  else:
-    hypernymWordA = {word[0]: word[1] for word in synsetA.hypernym_distances()}
-    hypernymWordB = {word[0]: word[1] for word in synsetB.hypernym_distances()}
-    commonValue = set(hypernymWordA.keys()).intersection(set(hypernymWordB.keys()))
-    if len(commonValue) <=0:
-      maxLength = 0
-    else:
-      distances = []
-      for common in commonValue:
-        commonWordADist = 0
-        commonWordBDist = 0
-        if common in hypernymWordA:
-          commonWordADist = hypernymWordA[common]
-        if common in hypernymWordB:
-          commonWordBDist = hypernymWordB[common]
-        maxDistance = max(commonWordADist, commonWordBDist)
-        distances.append(maxDistance)
-      maxLength = max(distances)
-  treeDist = (math.exp(smoothingFactor*maxLength)- math.exp(-smoothingFactor*maxLength))/(math.exp(smoothingFactor*maxLength) + math.exp(-smoothingFactor*maxLength))
-
-  return treeDist
 
 #################################################
 
@@ -403,47 +299,6 @@ def wordsSimilarity(wordA, wordB, method='semantic_similarity_synsets'):
 # failure fail 0.125
 # replacement replace 0.1111111111111111
 # leak leakage 1.0
-
-def synsetsSimilarity(synsetA, synsetB, method='semantic_similarity_synsets', disambiguation=True):
-  """
-    Compute synsets similarity
-    @ In, synsetA, wordnet.synset, the first synset
-    @ In, synsetB, wordnet.synset, the second synset
-    @ In, method, str, the method used to compute synset similarity
-      one of ['semantic_similarity_synsets', 'path', 'wup', 'lch', 'res', 'jcn', 'lin']
-    @ In, disambiguation, bool, True if disambiguation has been already performed
-    @ Out, similarity, float, [0, 1], the similarity score
-  """
-  method = method.lower()
-  if method != 'semantic_similarity_synsets' and not method.endswith('_similarity'):
-    method = '_'.join([method, 'similarity'])
-  wordnetSimMethod = ["path_similarity", "wup_similarity", "lch_similarity", "res_similarity", "jcn_similarity", "lin_similarity"]
-  sematicSimMethod = ['semantic_similarity_synsets']
-  synsetA = wn.synset(synsetA.name())
-  synsetB = wn.synset(synsetB.name())
-  if method in wordnetSimMethod:
-    if method in ["path_similarity", "wup_similarity"]:
-      similarity = getattr(wn, method)(synsetA, synsetB)
-    elif method == "lch_similarity":
-      if synsetA.name().split(".")[1] == synsetB.name().split(".")[1]:
-        similarity = getattr(wn, method)(synsetA, synsetB)
-      else:
-        similarity = 0.0
-    else:
-      brownIc = wordnet_ic.ic('ic-brown.dat')
-      if synsetA.name().split(".")[1] == synsetB.name().split(".")[1]:
-        try:
-          similarity = getattr(wn, method)(synsetA, synsetB, brownIc)
-        except:
-          similarity = 0.0
-      else:
-        similarity = 0.0
-  elif method in sematicSimMethod:
-    similarity = semanticSimilaritySynsets(synsetA, synsetB, disambiguation=disambiguation)
-  else:
-    raise ValueError(f'{method} is not valid, please use one of {wordnetSimMethod+sematicSimMethod}')
-
-  return similarity
 
 
 def wordSenseDisambiguation(word, sentence, senseMethod='simple_lesk', simMethod='path'):
@@ -553,49 +408,6 @@ def sentenceSimilarityWithDisambiguation(sentenceA, sentenceB, senseMethod='simp
   similarity = delta * semanticSimilarityUsingDisambiguatedSynsets(synsetsA, synsetsB, simMethod=simMethod) + (1.0-delta)* wordOrderSimilaritySentences(sentenceA, sentenceB)
   return similarity
 
-def semanticSimilarityUsingDisambiguatedSynsets(synsetsA, synsetsB, simMethod='semantic_similarity_synsets'):
-  """
-    Compute semantic similarity for given synsets while disambiguation has been already performed for given synsets
-    @ In, synsetsA, set/list, list of synsets
-    @ In, synsetsB, set/list, list of synsets
-    @ In, simMethod, str, method for similarity analysis in the construction of semantic vectors
-      one of ['semantic_similarity_synsets', 'path', 'wup', 'lch', 'res', 'jcn', 'lin']
-    @ Out, semSimilarity, float, [0, 1], the similarity score
-  """
-  jointWordSynsets = set(synsetsA).union(set(synsetsB))
-  wordVectorA = constructSemanticVectorUsingDisambiguatedSynsets(synsetsA, jointWordSynsets, simMethod=simMethod)
-  wordVectorB = constructSemanticVectorUsingDisambiguatedSynsets(synsetsB, jointWordSynsets, simMethod=simMethod)
-  semSimilarity = np.dot(wordVectorA, wordVectorB)/(np.linalg.norm(wordVectorA)*np.linalg.norm(wordVectorB))
-  return semSimilarity
-
-def constructSemanticVectorUsingDisambiguatedSynsets(wordSynsets, jointWordSynsets, simMethod='semantic_similarity_synsets'):
-  """
-    Construct semantic vector while disambiguation has been already performed
-    @ In, wordSynsets, set/list, set of words synsets
-    @ In, jointWords, set, set of joint words synsets
-    @ In, simMethod, str, method for similarity analysis in the construction of semantic vectors
-      one of ['semantic_similarity_synsets', 'path', 'wup', 'lch', 'res', 'jcn', 'lin']
-    @ Out, vector, numpy.array, semantic vector with disambiguation
-  """
-  wordSynsets = set(wordSynsets)
-  vector = np.zeros(len(jointWordSynsets))
-  for i, jointSynset in enumerate(jointWordSynsets):
-    simVector = []
-    if jointSynset in wordSynsets:
-      vector[i] = 1
-    else:
-      for synsetB in wordSynsets:
-        similarity = synsetsSimilarity(jointSynset, synsetB, method=simMethod, disambiguation=True)
-        simVector.append(similarity)
-      maxSim = max(simVector)
-      # if similarity < 0.2, treat it as noise and reset it to 0.0
-      if maxSim >= 0.2:
-        vector[i] = maxSim
-      else:
-        vector[i] = 0.0
-  return vector
-
-
 def convertSentsToSynsetsWithDisambiguation(sentList):
   """
     Use sentence itself to identify the best synset
@@ -606,3 +418,43 @@ def convertSentsToSynsetsWithDisambiguation(sentList):
     bestSyn = [wn.synset(syn.name()) for syn in bestSyn if syn is not None]
     sentSynsets.append(bestSyn)
   return sentSynsets
+
+def convertToSynsets(wordSet):
+  # keep the order (works for python3.7+)
+  wordList = list(dict.fromkeys(wordSet))
+  synsets = [list(wn.synsets(word)) for word in wordList]
+  return wordList, synsets
+
+def identifyBestSynset(word, jointWordList, jointSynsetList):
+  wordList = copy.copy(jointWordList)
+  synsets = copy.copy(jointSynsetList)
+  if word in wordList:
+    i = wordList.index(word)
+    wordList.remove(word)
+    synsets.remove(synsets[i])
+  synsets = [item for sub in synsets for item in sub]
+  wordSyns = wn.synsets(word)
+  if len(wordSyns) == 0:
+    return None
+  max = -1
+  bestSyn = wordSyns[0]
+  for syn in wordSyns:
+    similarity = [wn.path_similarity(syn, syn1) for syn1 in synsets]
+    temp = np.max(similarity)
+    if temp > max:
+      bestSyn = syn
+      max = temp
+  return bestSyn
+
+def convertSentsToSynsets(sentList):
+  """
+    Use sentence itself to identify the best synset
+  """
+  sentSynsets = []
+  for sent in sentList:
+    wordList, synsetsList = convertToSynsets([e.strip() for e in sent.split()])
+    bestSyn = [identifyBestSynset(word, wordList, synsetsList) for word in wordList]
+    bestSyn = list(filter(None, bestSyn))
+    sentSynsets.append(bestSyn)
+  return sentSynsets
+
